+++ crates/bonsaidb/src/any_connection.rs
@@ -246,7 +246,7 @@ impl<B: Backend> Connection for AnyDatabase<B> {
         &self,
         ids: R,
         order: Sort,
-        limit: Option<usize>,
+        limit: Option<u32>,
     ) -> Result<Vec<OwnedDocument>, bonsaidb_core::Error>
     where
         C: Collection,
@@ -275,7 +275,7 @@ impl<B: Backend> Connection for AnyDatabase<B> {
         &self,
         key: Option<QueryKey<V::Key>>,
         order: Sort,
-        limit: Option<usize>,
+        limit: Option<u32>,
         access_policy: AccessPolicy,
     ) -> Result<Vec<Map<V::Key, V::Value>>, bonsaidb_core::Error>
     where
@@ -291,7 +291,7 @@ impl<B: Backend> Connection for AnyDatabase<B> {
         &self,
         key: Option<QueryKey<V::Key>>,
         order: Sort,
-        limit: Option<usize>,
+        limit: Option<u32>,
         access_policy: AccessPolicy,
     ) -> Result<MappedDocuments<OwnedDocument, V>, bonsaidb_core::Error>
     where
@@ -366,7 +366,7 @@ impl<B: Backend> Connection for AnyDatabase<B> {
     async fn list_executed_transactions(
         &self,
         starting_id: Option<u64>,
-        result_limit: Option<usize>,
+        result_limit: Option<u32>,
     ) -> Result<Vec<Executed>, bonsaidb_core::Error> {
         match self {
             Self::Local(server) => {
